pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yuno-miles-coder"
version = "0.1.0"
description = "yuno-miles-coder: chaotic meme-style code explanations (Yuno Miles energy)."
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"

dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
  "pyarrow>=15.0.0",
  "fastapi>=0.109.0",
  "uvicorn>=0.27.0",
]

[project.scripts]
yuno-miles-coder = "app.cli:main"

[tool.setuptools]
packages = ["app"]
